Break qualification controller import cycle
